Quick commit

+++ routes/web.php
@@ -24,15 +24,15 @@ Route::middleware('auth')
         //==============================================================
         // ROTAS PARA USUÁRIOS  
 
-        Route::get('/users/search', [UserController::class, 'search'])->name('users.search');
-        Route::delete('/users/{user}/destroy', [UserController::class, 'destroy'])->name('users.destroy');
-        Route::post('/users/{id}/upload', [UserController::class, 'upload'])->name('users.upload');
-        Route::get('/users/create',[UserController::class, 'create'])->name('users.create');
-        Route::get('/users/{user}', [UserController::class, 'show'])->name('users.show');
-        Route::put('/users/{user}',[UserController::class, 'update'])->name('users.update');
-        Route::get('/users/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
-        Route::post('/users',[UserController::class, 'store'])->name('users.store');
-        Route::get('/users',[UserController::class, 'index'])->name('users.index');
+        Route::get('/users/search', [UserController::class, 'search'])->name('users.search')->middleware(checkpermission::class);
+        Route::delete('/users/{user}/destroy', [UserController::class, 'destroy'])->name('users.destroy')->middleware(checkpermission::class);
+        Route::post('/users/{id}/upload', [UserController::class, 'upload'])->name('users.upload')->middleware(checkpermission::class);
+        Route::get('/users/create',[UserController::class, 'create'])->name('users.create')->middleware(checkpermission::class);
+        Route::get('/users/{user}', [UserController::class, 'show'])->name('users.show')->middleware(checkpermission::class);
+        Route::put('/users/{user}',[UserController::class, 'update'])->name('users.update')->middleware(checkpermission::class);
+        Route::get('/users/{user}/edit', [UserController::class, 'edit'])->name('users.edit')->middleware(checkpermission::class);
+        Route::post('/users',[UserController::class, 'store'])->name('users.store')->middleware(checkpermission::class);
+        Route::get('/users',[UserController::class, 'index'])->name('users.index')->middleware(checkpermission::class);
 
         //======================================================
         // ROTAS PARA EMPRESA INUSITTA
@@ -983,26 +983,27 @@ Route::middleware('auth')
 //==============================================================
 // ROTAS PARA FILES
 
-
-Route::post('/files/{id}/ativar', [FileController::class, 'ativar'])->name('files.ativar');
-Route::post('/files/{id}/inativar', [FileController::class, 'inativar'])->name('files.inativar');
-Route::post('/files/{id}/reprovar', [FileController::class, 'reprovar'])->name('files.reprovar');
-Route::post('/files/{id}/aprovar', [FileController::class, 'aprovar'])->name('files.aprovar');
-Route::get('/files/arquivosativos', [FileController::class, 'arquivosAtivos'])->name('arquivosAtivos.index');
-Route::get('/files/arquivosreprovados',[FileController::class, 'arquivosReprovados'])->name('arquivosReprovados.index');
-Route::get('/files/arquivosaprovados', [FileController::class, 'arquivosAprovados'])->name('arquivosAprovados.index');
-Route::get('/files/arquivosinativos', [FileController::class, 'arquivosInativos'])->name('arquivosInativos.index');
-Route::get('/files/aguardandoaprovacao', [FileController::class, 'aguardandoAprovacao'])->name('aguardandoAprovacao.index');
-Route::get('/files/search', [FileController::class, 'search'])->name('files.search');
+Route::middleware(checkPermission::class)->group(function(){
+    Route::post('/files/{id}/ativar', [FileController::class, 'ativar'])->name('files.ativar');
+    Route::post('/files/{id}/inativar', [FileController::class, 'inativar'])->name('files.inativar');
+    Route::post('/files/{id}/reprovar', [FileController::class, 'reprovar'])->name('files.reprovar');
+    Route::post('/files/{id}/aprovar', [FileController::class, 'aprovar'])->name('files.aprovar');
+    Route::get('/files/arquivosativos', [FileController::class, 'arquivosAtivos'])->name('arquivosAtivos.index');
+    Route::get('/files/arquivosreprovados',[FileController::class, 'arquivosReprovados'])->name('arquivosReprovados.index');
+    Route::get('/files/arquivosaprovados', [FileController::class, 'arquivosAprovados'])->name('arquivosAprovados.index');
+    Route::get('/files/arquivosinativos', [FileController::class, 'arquivosInativos'])->name('arquivosInativos.index');
+    Route::get('/files/aguardandoaprovacao', [FileController::class, 'aguardandoAprovacao'])->name('aguardandoAprovacao.index');
+    Route::get('/files/search', [FileController::class, 'search'])->name('files.search');
+    Route::delete('/files/{file}/destroy', [FileController::class, 'destroy'])->name('files.destroy');
+    Route::post('/files/{id}/upload', [FileController::class, 'upload'])->name('files.upload');
+    Route::get('/files/create',[FileController::class, 'create'])->name('files.create');
+    Route::get('/files/{file}', [FileController::class, 'show'])->name('files.show');
+    Route::put('/files/{file}',[FileController::class, 'update'])->name('files.update');
+    Route::get('/files/{file}/edit', [FileController::class, 'edit'])->name('files.edit');
+    Route::post('/files',[FileController::class, 'store'])->name('files.store');
+    Route::get('/files',[FileController::class, 'index'])->name('files.index');
+});
 Route::get('/files/searchCommonUser', [FileController::class, 'searchCommonUser'])->name('filesCommonUser.search');
-Route::delete('/files/{file}/destroy', [FileController::class, 'destroy'])->name('files.destroy');
-Route::post('/files/{id}/upload', [FileController::class, 'upload'])->name('files.upload');
-Route::get('/files/create',[FileController::class, 'create'])->name('files.create');
-Route::get('/files/{file}', [FileController::class, 'show'])->name('files.show');
-Route::put('/files/{file}',[FileController::class, 'update'])->name('files.update');
-Route::get('/files/{file}/edit', [FileController::class, 'edit'])->name('files.edit');
-Route::post('/files',[FileController::class, 'store'])->name('files.store');
-Route::get('/files',[FileController::class, 'index'])->name('files.index')->middleware(checkPermission::class);
 Route::get('/files/view/{id}', [FileController::class, 'view'])->name('files.view')->middleware('auth');
 
 //==============================================================
@@ -1011,34 +1012,36 @@ Route::get('/files/view/{id}', [FileController::class, 'view'])->name('files.vie
 //==============================================================
 // ROTAS PARA EMPRESAS
 
-Route::delete('/companies/{company}/destroy', [CompanyController::class, 'destroy'])->name('companies.destroy');
-Route::get('/companies/create', [CompanyController::class, 'create'])->name('companies.create');
-Route::get('/companies/{company}', [CompanyController::class, 'show'])->name('companies.show');
-Route::put('/companies/{company}',[CompanyController::class, 'update'])->name('companies.update');
-Route::get('/companies/{company}/edit', [CompanyController::class, 'edit'])->name('companies.edit');
-Route::post('/companies',[CompanyController::class, 'store'])->name('companies.store');
-Route::get('/companies',[CompanyController::class, 'index'])->name('companies.index');
-
+Route::middleware([checkPermission::class])->group(function(){
+    Route::delete('/companies/{company}/destroy', [CompanyController::class, 'destroy'])->name('companies.destroy');
+    Route::get('/companies/create', [CompanyController::class, 'create'])->name('companies.create');
+    Route::get('/companies/{company}', [CompanyController::class, 'show'])->name('companies.show');
+    Route::put('/companies/{company}',[CompanyController::class, 'update'])->name('companies.update');
+    Route::get('/companies/{company}/edit', [CompanyController::class, 'edit'])->name('companies.edit');
+    Route::post('/companies',[CompanyController::class, 'store'])->name('companies.store');
+    Route::get('/companies',[CompanyController::class, 'index'])->name('companies.index');
+});
 //==============================================================
 
 
 //==============================================================
 // ROTAS PARA SETORES
 
-Route::delete('/sectors/{sector}/destroy', [SectorController::class, 'destroy'])->name('sectors.destroy');
-Route::get('/sectors/create', [SectorController::class, 'create'])->name('sectors.create');
-Route::get('/sectors/{sector}', [SectorController::class, 'show'])->name('sectors.show');
-Route::put('/sectors/{sector}',[SectorController::class, 'update'])->name('sectors.update');
-Route::get('/sectors/{sector}/edit', [SectorController::class, 'edit'])->name('sectors.edit');
-Route::post('/sectors',[SectorController::class, 'store'])->name('sectors.store');
-Route::get('/sectors',[SectorController::class, 'index'])->name('sectors.index');
-
+Route::middleware([checkPermission::class])->group(function(){
+    Route::delete('/sectors/{sector}/destroy', [SectorController::class, 'destroy'])->name('sectors.destroy');
+    Route::get('/sectors/create', [SectorController::class, 'create'])->name('sectors.create');
+    Route::get('/sectors/{sector}', [SectorController::class, 'show'])->name('sectors.show');
+    Route::put('/sectors/{sector}',[SectorController::class, 'update'])->name('sectors.update');
+    Route::get('/sectors/{sector}/edit', [SectorController::class, 'edit'])->name('sectors.edit');
+    Route::post('/sectors',[SectorController::class, 'store'])->name('sectors.store');
+    Route::get('/sectors',[SectorController::class, 'index'])->name('sectors.index');
+});
 //==============================================================
 
 //==============================================================
 // ROTAS PARA MACROS
 
-Route::middleware('auth')
+Route::middleware(['auth', checkPermission::class])
     ->prefix('admin')
     ->group(function(){
         
@@ -1061,14 +1064,15 @@ Route::middleware('auth')
 //==============================================================
 // ROTAS PARA PERMISSÕES
 
-Route::delete('/permissions/{permission}/destroy', [PermissionController::class, 'destroy'])->name('permissions.destroy');
-Route::get('/permissions/create', [PermissionController::class, 'create'])->name('permissions.create');
-Route::get('/permissions/{permission}', [PermissionController::class, 'show'])->name('permissions.show');
-Route::put('/permissions/{permission}',[PermissionController::class, 'update'])->name('permissions.update');
-Route::get('/permissions/{permission}/edit', [PermissionController::class, 'edit'])->name('permissions.edit');
-Route::post('/permissions',[PermissionController::class, 'store'])->name('permissions.store');
-Route::get('/permissions',[PermissionController::class, 'index'])->name('permissions.index');
-
+Route::middleware([checkPermission::class])->group(function(){
+    Route::delete('/permissions/{permission}/destroy', [PermissionController::class, 'destroy'])->name('permissions.destroy');
+    Route::get('/permissions/create', [PermissionController::class, 'create'])->name('permissions.create');
+    Route::get('/permissions/{permission}', [PermissionController::class, 'show'])->name('permissions.show');
+    Route::put('/permissions/{permission}',[PermissionController::class, 'update'])->name('permissions.update');
+    Route::get('/permissions/{permission}/edit', [PermissionController::class, 'edit'])->name('permissions.edit');
+    Route::post('/permissions',[PermissionController::class, 'store'])->name('permissions.store');
+    Route::get('/permissions',[PermissionController::class, 'index'])->name('permissions.index');
+});
 //==============================================================
 
 
@@ -1089,7 +1093,7 @@ Route::get('/dashboard', function () {
     return view('dashboard');
 })->middleware(['auth', 'verified'])->name('dashboard');
 
-Route::middleware('auth')->group(function () {
+Route::middleware(['auth', checkPermission::class])->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
     Route::patch('/profile', [ProfileController::class, 'update'])->name('profile.update');
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
